requirements.txt
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
oauth2client
aiohttp
pytz
python-telegram-bot
python-dotenv
httpx
openai
